Refactorize
