+++ 0_tutorial_cgraph/app.cpp
@@ -3,15 +3,6 @@
 #include <numeric>
 #include <taichi/cpp/taichi.hpp>
 
-namespace {
-void check_taichi_error(const std::string& msg) {
-  TiError error = ti_get_last_error(0, nullptr);
-  if (error < TI_ERROR_SUCCESS) {
-    throw std::runtime_error(msg);
-  }
-}
-}
-
 struct App0_tutorial {
   // This is different from what used in python script since compiled shaders are compatible with dynamic ndarray shape
   static const uint32_t NPARTICLE = 8192 * 2;
@@ -25,11 +16,11 @@ struct App0_tutorial {
   App0_tutorial() {
     runtime_ = ti::Runtime(TI_ARCH_VULKAN);
     module_ = runtime_.load_aot_module("0_tutorial_cgraph/assets/tutorial");
-    check_taichi_error("load_aot_module failed");
+    ti::check_last_error();
     g_demo_ = module_.get_compute_graph("demo_graph");
-    check_taichi_error("get_compute_graph failed");
+    ti::check_last_error();
     x_ = runtime_.allocate_ndarray<float>({NPARTICLE}, {}, /*host_accessible=*/true);
-    check_taichi_error("allocate_ndarray failed");
+    ti::check_last_error();
     std::cout << "Initialized!" << std::endl;
   }
 
@@ -41,7 +32,7 @@ struct App0_tutorial {
 
     g_demo_.launch();
     runtime_.wait();
-    check_taichi_error("cgraph launch failed");
+    ti::check_last_error();
 
     std::vector<float> dst(NPARTICLE);
     x_.read(dst);
