+++ 2_mpm88/app.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include "glm/glm.hpp"
 #include "taichi/aot_demo/framework.hpp"
-#include "taichi/aot_demo/interop/cross_device_copy.hpp"
+#include "taichi/aot_demo/vulkan/interop/cross_device_copy.hpp"
 
 using namespace ti::aot_demo;
 
@@ -104,7 +104,7 @@ struct App2_mpm88 : public App {
     }
 
     // 2. Load AOT module
-#ifdef TI_AOT_DEMO_WITH_ANDROID_APP
+#ifdef TI_AOT_DEMO_ANDROID_APP
     std::vector<uint8_t> tcm;
     F_->asset_mgr().load_file("E2_mpm88.tcm", tcm);
     module_ = runtime_.create_aot_module(tcm);
