+++ 3_implicit_fem/app.cpp
@@ -55,7 +55,7 @@ struct App3_implicit_fem : public App {
     GraphicsRuntime& runtime = F_->runtime();
     Renderer& renderer = F_->renderer();
 
-#ifdef TI_AOT_DEMO_WITH_ANDROID_APP
+#ifdef TI_AOT_DEMO_ANDROID_APP
     std::vector<uint8_t> tcm;
     F_->asset_mgr().load_file("E3_implicit_fem.tcm", tcm);
     module_ = runtime.create_aot_module(tcm);
