+++ framework/include/taichi/aot_demo/common.hpp
@@ -10,6 +10,27 @@
 
 #include "glm/glm.hpp"
 
+#if TI_WITH_VULKAN
+#ifdef _WIN32
+#define VK_USE_PLATFORM_WIN32_KHR
+#endif  // _WIN32
+#ifdef ANDROID
+#define VK_USE_PLATFORM_ANDROID_KHR 1
+#endif // ANDROID
+#include <vulkan/vulkan.h>
+#define VMA_DYNAMIC_VULKAN_FUNCTIONS
+#include <vk_mem_alloc.h>
+#endif  // TI_WITH_VULKAN
+
+#ifdef TI_WITH_METAL
+#ifdef __OBJC__
+#import <Foundation/Foundation.h>
+#import <QuartzCore/CAMetalLayer.h>
+#import <Metal/Metal.h>
+#import <MetalKit/MTKView.h>
+#endif  // __OBJC__
+#endif  // TI_WITH_METAL
+
 #ifdef TI_WITH_CUDA
 #include <cuda.h>
 #endif
@@ -18,21 +39,11 @@
 #include "glad/gl.h"
 #endif // TI_WITH_OPENGL
 
-#if TI_AOT_DEMO_ANDROID_APP
-#define VK_USE_PLATFORM_ANDROID_KHR 1
-#endif // TI_AOT_DEMO_ANDROID_APP
-
-#ifdef _WIN32
-#define VK_USE_PLATFORM_WIN32_KHR
-#endif // _WIN32
-#include <vulkan/vulkan.h>
 #include "taichi/cpp/taichi.hpp"
-#define VMA_DYNAMIC_VULKAN_FUNCTIONS
-#include <vk_mem_alloc.h>
 
-#if TI_AOT_DEMO_WITH_GLFW
+#if TI_AOT_DEMO_GLFW
 #include "GLFW/glfw3.h"
-#endif // TI_AOT_DEMO_WITH_GLFW
+#endif // TI_AOT_DEMO_GLFW
 
 #if TI_AOT_DEMO_ANDROID_APP
 #include <android/native_window.h>
