+++ framework/src/taichi/aot_demo/vulkan/interop/cross_device_copy.cpp
@@ -1,6 +1,6 @@
-#include "taichi/aot_demo/interop/cross_device_copy.hpp"
-#include "taichi/aot_demo/interop/common_utils.hpp"
-#include "taichi/aot_demo/renderer.hpp"
+#include "taichi/aot_demo/vulkan/interop/cross_device_copy.hpp"
+#include "taichi/aot_demo/vulkan/interop/common_utils.hpp"
+#include "taichi/aot_demo/vulkan/vulkan_renderer.hpp"
 
 
 namespace ti {
